filter and jobs in same page
