+++ src/app/components/provider/job-card/job-card.component.html
@@ -2,7 +2,10 @@
     <p class="title-container">{{jobTitle}}</p>
     <p class="description-container">{{jobDescription}}</p>
     <div class="bottom-container">
-        <p *ngIf="jobStatus == 'AVAILABLE'"class="status">Disponibil</p>
-        <p-button (click)="onBtnClick()"class="assign-btn" label="Asigneaza Job"></p-button>
+        <div class="pContainer">
+            <p class="pCategory">{{jobCategory}}</p>
+            <p *ngIf="jobStatus == 'AVAILABLE'" class="status">Disponibil</p>
+        </div>
+        <p-button (click)="onBtnClick()" class="assign-btn" label="Asigneaza Job"></p-button>
     </div>
 </div>
