+++ src/app/components/provider/provider-jobs/provider-jobs.component.ts
@@ -1,6 +1,6 @@
 import {Component, OnInit, ViewEncapsulation} from '@angular/core';
 import {JobService} from 'src/app/services/job.service';
-import {Job} from '../../shared/model';
+import {Job, Category} from '../../shared/model';
 import {Observable} from 'rxjs';
 import {MessageService} from 'primeng/api';
 import {UserService} from '../../../services/user.service';
@@ -17,6 +17,9 @@ export class ProviderJobsComponent implements OnInit {
   protected availableJobs: Job[] = [];
   protected userId: string = '';
 
+  categories: any[];
+  selectedCategories: Category[] = [];
+
   private limit: number = 1000;
   private pageNumber: number = 0;
 
@@ -27,16 +30,49 @@ export class ProviderJobsComponent implements OnInit {
 
   ngOnInit() {
     this.userId = this.loginService.getUser().id;
-    this.loadData().subscribe((jobs) => this.availableJobs = jobs.filter(job => job.status === 'AVAILABLE'));
+    this.loadData();
+    this.categories = [];
 
+    const enumObject = Object(Category);
+    for (const key in enumObject) {
+      this.categories.push({'name': enumObject[key]});
+    }
   }
 
-  private loadData(): Observable<Job[]> {
-    return this.jobService.getAllJobs();
+  private loadData() {
+    this.jobService.getAllJobs().subscribe((jobs) => {
+      this.availableJobs = jobs.filter(job => job.status === 'AVAILABLE')
+    },(error) => {
+      this.messageService.add({severity: 'error', summary: 'Error', detail: 'Datele nu au putut fi incarcate, incearca din nou'});
+    });
   }
 
   private onClickAssign(event: string): void {
     this.jobService.assignJob(this.userId, event);
     this.messageService.add({severity: 'success', summary: 'Succes', detail: 'Job-ul a fost asignat'});
+  }
+
+
+  onFilter() {
+    const filterCategories = [];
+
+    for (var i = 0; i < this.selectedCategories.length; i++) {
+      filterCategories.push(this.selectedCategories[i]['name']);
+    }
+    console.log(`Reset ${filterCategories}`);
+    // no filters, just load data
+    if (filterCategories == null || filterCategories.length == 0) {
+      this.loadData()
+    } else {
+      this.jobService.getFilteredJobs(filterCategories).subscribe((jobs) => {
+        this.availableJobs = jobs;
+      }, (error) => {
+        this.messageService.add({severity: 'error', summary: 'Error', detail: 'Datele nu au putut fi incarcate, incearca din nou'});
+      });
+    }
+  }
+
+  onReset() {
+    this.selectedCategories = [];
   }
 }
