+++ src/app/components/provider/provider-routing.module.ts
@@ -6,7 +6,6 @@ import {ProviderSettingsComponent} from './provider-settings/provider-settings.c
 import { NewReviewComponent } from './new-review/new-review.component';
 import {AddedReviewListComponent} from './added-review-list/added-review-list.component';
 import {ReceivedReviewListComponent} from './received-review-list/received-review-list.component';
-import {ProviderSearchComponent} from './provider-search/provider-search.component';
 import { ProviderJobsComponent } from './provider-jobs/provider-jobs.component';
 import {StatisticsComponent} from '../shared/statistics/statistics.component';
 
@@ -18,10 +17,6 @@ const routes = [
   {
     path: 'provider/request',
     component: RequestListComponent
-  },
-  {
-    path: 'provider/search',
-    component: ProviderSearchComponent
   },
   {
     path: 'provider/account-settings',
