--- src/app/components/provider/provider-search/provider-search.component.css
@@ -1,84 +1,0 @@
-.my-div-class {
-  margin: auto;
-  width: 550px;
-  background: white;
-}
-
-.titlu {
-  margin: auto;
-}
-
-.my-border {
-  border-radius: 10px;
-  border: 0.5px solid #eaeeef;
-}
-
-.my-inputs {
-  padding: 12px;
-  border: 1px solid #CCCCCC;
-  width: 100%;
-}
-
-.my-select {
-  padding: 12px;
-  border: 1px solid #CCCCCC;
-  width: 100%;
-}
-
-input[type=submit] {
-  color: white;
-  background: #396dce;
-}
-
-input[type=submit]:hover {
-  color: white;
-  transition: 0.5s;
-  background: #1c3667;
-}
-
-::-webkit-input-placeholder { /* Chrome/Opera/Safari */
-  color: #d1d1d1;
-}
-
-::-moz-placeholder { /* Firefox 19+ */
-  color: #d1d1d1;
-}
-
-:-ms-input-placeholder { /* IE 10+ */
-  color: #d1d1d1;
-}
-
-:-moz-placeholder { /* Firefox 18- */
-  color: #d1d1d1;
-}
-
-.float-container {
-  border: solid 1px #ccc;
-  position: relative;
-
-}
-
-.float-container input {
-  border: none;
-  outline: 0;
-  font-size: 16px;
-  padding: 20px 0 10px 12px;
-}
-
-label {
-  color: #ccc;
-  padding-left: 12px;
-  padding-top: 2px;
-  font-size: 14px;
-  position: absolute;
-}
-
-.float-container > input.ng-invalid {
-  border-left: 3px solid #FF5D5D !important;
-}
-
-
-p-button {
-  position: center;
-  width: inherit;
-}
