--- src/app/components/provider/provider-search/provider-search.component.html
@@ -1,48 +1,0 @@
-<app-provider-toolbar></app-provider-toolbar>
-<p-toast position="top-right"></p-toast>
-
-<div class="container-fluid my-div-class">
-  <div class="row mb-3 mt-3">
-    <div class="col text-center">
-      <h3 class="text-center">
-        Caută Joburi
-      </h3>
-    </div>
-  </div>
-</div>
-
-
-
-<div class="container-fluid my-border mb-4">
-      <div class="text-center">
-          Categorie de căutare:
-      </div>
-</div>
-
-<div class="row" style="justify-content:center">
-  <div>
-    <p-multiSelect optionLabel="name" [options]="categories" [(ngModel)]="selectedCategories" [panelStyle]="{minWidth:'16em'}"></p-multiSelect>
-  </div>
-</div>
-<br>
-<div class="row" style="justify-content:center">
-  <div >
-    <p-button label="Reset" (onClick)="onReset()"></p-button>
-    <p-button label="Filter" (onClick)="onFilter()"></p-button>
-  </div>
-
-  <br><br>
-
-  <div class="card-jobs-container">
-    <div class="card-container" *ngFor="let job of jobsFound">
-        <app-job-card
-            (assignedClick)="inClickAssign($event)"
-            [jobTitle]="job.title"
-            [jobDescription]="job.description"
-            [jobId]="job.id"
-            [jobStatus]="job.status">
-
-        </app-job-card>
-    </div>
-</div>
-</div>
